:truck: - Move icons to assets folder
